pyproject.toml
[build-system]
requires = ["setuptools>=61.0",
            "platformdirs>=2.5.2",
            "PyYAML>=6.0",
            ]
build-backend = "setuptools.build_meta"

[project]
name = "combine_settings"
version = "1.1.0"
authors = [
  { name="David Willcox", email="daw30410@yahoo.com" },
]
description = "Combine application settings from multiple files"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
                "platformdirs>=2.5.2",
                "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/dawillcox/python_combine_settings"
"Bug Tracker" = "https://github.com/dawillcox/python_combine_settings/issues"
Documentation = "https://github.com/dawillcox/python_combine_settings/blob/main/README.md"
Repository = "https://github.com//dawillcox/python_combine_settings.git"
Changelog = "https://github.com/dawillcox/python_combine_settings/blob/main/CHANGELOG.md"

